[iOS] Minor cleanup for test formatting

+++ sdk/iOS/test/MSTableTests.m
@@ -40,6 +40,7 @@
 
 #pragma mark * Init Method Tests
 
+
 -(void) testInitWithNameAndClient
 {
     MSTable *table = [[MSTable alloc] initWithName:@"SomeName" client:client];
@@ -63,6 +64,7 @@
 
 
 #pragma mark * Insert Method Tests
+
 
 // See the WindowsAzureMobileServicesFunctionalTests.m tests for additional
 // insert tests that require a working Microsoft Azure Mobile Service.
@@ -416,6 +418,7 @@
     
     XCTAssertTrue([self waitForTest:0.1], @"Test timed out.");
 }
+
 
 #pragma mark * Update Method Tests
 
@@ -1678,7 +1681,9 @@
     XCTAssertTrue([self waitForTest:0.1], @"Test timed out.");
 }
 
-# pragma mark System Property Tests
+
+#pragma mark System Property Tests
+
 
 -(void) testInsertStringIdPropertiesNotRemovedFromRequest
 {
@@ -1981,7 +1986,7 @@
         NSDictionary *item = @{@"id":@"an id",@"String":@"what?"};
         testFilter.dataToUse = [@"{\"id\":\"an id\",\"String\":\"Hey\"}" dataUsingEncoding:NSUTF8StringEncoding];
         [todoTable insert:item completion:^(NSDictionary *item, NSError *error) {
-            XCTAssertTrue([self checkRequestURL:[actualRequest URL] SystemProperty:systemProperty], @"Error for property %d with url: %@", systemProperty, [[actualRequest URL] query]);
+            XCTAssertTrue([self checkRequestURL:[actualRequest URL] SystemProperty:systemProperty], @"Error for property %lu with url: %@", (unsigned long)systemProperty, [[actualRequest URL] query]);
             done = YES;
         }];
         XCTAssertTrue([self waitForTest:0.1], @"Test timed out.");
